Don't logout root, update other copies of scripts

+++ ee/server/service/embedded_scripts/linux_lock.sh
@@ -33,6 +33,7 @@ logged_in=$(users | tr ' ' '\n' | sort  | uniq)
 
 # Logout any non-passwd users
 for user in $logged_in; do
+    [ "$user" = "root" ] && continue
     echo "Logging out $user"
     pkill -KILL -u "$user"
 done
